Do not fetch public_channels, instead use userBoot endpoint to fetch
userBoot, as far as I understand, fetches all joined public
channels. So, instead of fetching all public channels, we can simply
use userBoot to fetch what we need. Of course, user won't be able to
join a public channel but that is not something you do everyday and
it's a good compromise to not get rate limited.