attempt to reduce load of slack-user-list-update on reconnect
This operation still rate limits the token for slack teams with many
users.
For example, Clojurians has about 29k users and the call only works if
the token has not been used for other intensive update operations.

Here I lock the reconnect slack-user-list-update so that it only runs
every 45 seconds.